Almost done with recursive reconstruction

+++ 07_simultaneousPostProcessAndRecursiveExecutionWithOverlayFS/example_OpenFOAM-2.4.x/E.1.reconstructFromOverlay.sh
@@ -1,4 +1,5 @@
 #!/bin/bash -l
+#SBATCH --job-name=reconstructFromOverlay
 #SBATCH --ntasks=4
 #SBATCH --mem=4G
 #SBATCH --ntasks-per-node=28
@@ -7,12 +8,13 @@
 #SBATCH --partition=workq
 #SBATCH --export=none
 
-#0. Initial settings:
+#========================
+echo '#0. Initial settings:'
 unset XDG_RUNTIME_DIR #To avoid some annoying warnings when using some containers
 : ${surnameTag:=""}
-#: ${surnameTag:="_1234567_0"}
 
-#1. Loading the container settings, case settings and auxiliary functions (order is important)
+#========================
+echo '#1. Loading the container settings, case settings and auxiliary functions (order is important)'
 source $SLURM_SUBMIT_DIR/imageSettingsSingularity.sh
 source $SLURM_SUBMIT_DIR/caseSettingsFoam.sh
 overlayFunctionsScript=$auxScriptsDir/ofContainersOverlayFunctions.sh
@@ -23,7 +25,8 @@ else
    echo "Exiting"; exit 1
 fi
 
-#2. Going into the case and creating the logs directory
+#========================
+echo '#2. Going into the case and creating the logs directory'
 if [ -d $caseDir ]; then
    cd $caseDir
    echo "pwd=$(pwd)"
@@ -31,23 +34,48 @@ else
    echo "For some reason, the case=$caseDir, does not exist"
    echo "Exiting"; exit 1
 fi
-logsDir=./logs/post
+logsDir=$caseDir/logs/post
 if ! [ -d $logsDir ]; then
    mkdir -p $logsDir
 fi
 
-#3. Reading OpenFOAM decomposeParDict settings
+#========================
+echo '#3. Reading OpenFOAM decomposeParDict settings'
 foam_numberOfSubdomains=$(grep "^numberOfSubdomains" ./system/decomposeParDict | tr -dc '0-9')
 
-#4. Create the reconstruction array, intended times to be reconstructed are set with the reconstructTimes var
-#These formats are the only accepted by function "generateReconstructArray" (check the function definition for further information)
-reconstructTimes="all"
-#reconstructTimes="-2"
-#reconstructTimes="+2"
-#reconstructTimes="20"
-#reconstructTimes="50,60,70,80,90"
-#reconstructTimes="0:10"
-#reconstructTimes="10:24"
+#========================
+echo '#4. Create the reconstruction directory and cd into it'
+: ${reconstructDir:="${caseDir}/reconstructDir/${SLURM_JOB_NAME}"}
+if ! [ -d $reconstructDir ]; then
+   mkdir -p $reconstructDir
+   ln -s ${caseDir}/system ${reconstructDir}/system
+   ln -s ${caseDir}/constant ${reconstructDir}/constant
+   ln -s ${caseDir}/overlayFSDir ${reconstructDir}/overlayFSDir
+   for ii in $(seq 0 $((foam_numberOfSubdomains - 1))); do
+      if ! [ -d ${reconstructDir}/bakDir/bak.processor${ii} ]; then
+         mkdir -p ${reconstructDir}/bakDir/bak.processor${ii} 
+      fi
+      ln -s ${caseDir}/bakDir/bak.processor${ii}/constant ${reconstructDir}/bakDir/bak.processor${ii}/constant
+   done
+fi
+cd $reconstructDir
+
+#========================
+echo '#5. Create reconstruction array (intended times to be reconstructed are set with $reconstructTimes)'
+#IMPORTANT: Note the use of : ${var:="value"} bash syntax. This allows to receive this variable
+#           from a parent script, or as an "exported" value when submitting this script.
+#           This is useful for recursive jobs.
+#ALSO: The formats inside the apostrophes ("format") are the only accepted by function:
+#      "generateReconstructArray" (check the function definition for further information)
+echo "AEG:E.1.A: reconstructTimes=$reconstructTimes"
+: ${reconstructTimes:="all"}
+echo "AEG:E.1.B: reconstructTimes=$reconstructTimes"
+#: ${reconstructTimes:="-2"}
+#: ${reconstructTimes:="+2"}
+#: ${reconstructTimes:="20"}
+#: ${reconstructTimes:="50,60,70,80,90"}
+#: ${reconstructTimes:="0:10"}
+#: ${reconstructTimes:="10:24"}
 unset arrayReconstruct #This global variable will be re-created in the function below
 generateReconstructArray "$reconstructTimes" $insideDir $surnameTag;success=$? #Calling fucntion to generate "arrayReconstruct"
 if [ $success -ne 0 ]; then
@@ -55,15 +83,15 @@ if [ $success -ne 0 ]; then
    echo "Exiting";exit 1
 fi
 
-#5. Point the soft links to the ./bakDir/bak.processor* directories
+#6. Point the soft links to the ./bakDir/bak.processor* directories
 pointToBak $foam_numberOfSubdomains;success=$? #Calling function to point towards the ./bakDir/bak.processors
 if [ $success -ne 0 ]; then
    echo "Failed creating the soft links"
    echo "Exiting";exit 1
 fi
 
-#6. Generate a list of the time directories already inside ./bakDir/bak.processor* to be preserved and not removed after reconstruction
-#   (The earlier and latest times will be preserved)
+#7. Define the directories to be preserved (and not removed after reconst) in the ./bakDir/bak.processor*
+##7.1 Generate a list of the time directories already inside ./bakDir/bak.processor*
 ls -dt ./bakDir/bak.processor0/[0-9]* | sed "s,./bakDir/bak.processor0/,," > bakTimes.$SLURM_JOBID
 sort -n bakTimes.$SLURM_JOBID -o bakTimesSorted.$SLURM_JOBID
 rm bakTimes.$SLURM_JOBID
@@ -79,25 +107,29 @@ nTimeBak=$i
 rm bakTimesSorted.$SLURM_JOBID
 keepTimesArr[0]=-1
 nKeepTimes=0
-if [ $nTimeBak -eq 0 ]; then
-   echo "NO time directories available in ./bakDir/bak.processor0"
-else
-   echo "The minTime already in bak is ${bakArr[0]}"
-   echo "The minTime to be reconstructed is ${arrayReconstruct[0]}"
-   if [ $(echo "${bakArr[0]} < ${arrayReconstruct[0]}" | bc -l) -eq 1 ]; then
-      keepTimesArr[$nKeepTimes]=${bakArr[0]};(( nKeepTimes++ )) 
+## Decide if we want to keep the first and last times in the ./bakDir/bak.processor* decomposed directories
+keepFirstAndLast=false
+if [ "$keepFirstAndLast" = "true" ]; then
+   if [ $nTimeBak -eq 0 ]; then
+      echo "NO time directories available in ./bakDir/bak.processor0"
    else
-      keepTimesArr[$nKeepTimes]=${arrayReconstruct[0]};(( nKeepTimes++ )) 
+      echo "The minTime already in bak is ${bakArr[0]}"
+      echo "The minTime to be reconstructed is ${arrayReconstruct[0]}"
+      if [ $(echo "${bakArr[0]} < ${arrayReconstruct[0]}" | bc -l) -eq 1 ]; then
+         keepTimesArr[$nKeepTimes]=${bakArr[0]};(( nKeepTimes++ )) 
+      else
+         keepTimesArr[$nKeepTimes]=${arrayReconstruct[0]};(( nKeepTimes++ )) 
+      fi
+      echo "Time ${keepTimesArr[$((nKeepTimes-1))]} will be preserved after reconstruction"
+      echo "The maxTime already in bak is ${bakArr[-1]}"
+      echo "The maxTime to be reconstructed is ${arrayReconstruct[-1]}"
+      if [ $(echo "${bakArr[-1]} > ${arrayReconstruct[-1]}" | bc -l) -eq 1 ]; then
+         keepTimesArr[$nKeepTimes]=${bakArr[-1]};(( nKeepTimes++ )) 
+      else
+         keepTimesArr[$nKeepTimes]=${arrayReconstruct[-1]};(( nKeepTimes++ )) 
+      fi
+      echo "Time ${keepTimesArr[$((nKeepTimes-1))]} will be preserved after reconstruction"
    fi
-   echo "Time ${keepTimesArr[$((nKeepTimes-1))]} will be preserved after reconstruction"
-   echo "The maxTime already in bak is ${bakArr[-1]}"
-   echo "The maxTime to be reconstructed is ${arrayReconstruct[-1]}"
-   if [ $(echo "${bakArr[-1]} > ${arrayReconstruct[-1]}" | bc -l) -eq 1 ]; then
-      keepTimesArr[$nKeepTimes]=${bakArr[-1]};(( nKeepTimes++ )) 
-   else
-      keepTimesArr[$nKeepTimes]=${arrayReconstruct[-1]};(( nKeepTimes++ )) 
-   fi
-   echo "Time ${keepTimesArr[$((nKeepTimes-1))]} will be preserved after reconstruction"
 fi
 ## Add other times to the list if desired
 #keepTimesArr[$nKeepTimes]=28;(( nKeepTimes++ ))
@@ -106,14 +138,16 @@ fi
 echo "All times to be preserved in ./bakDir/bak.processor* are:"
 echo "${keepTimesArr[@]}"
 
-#7. Check for already reconstructed cases and set the `-time $timeString` argument for the reconstructPar tool
+#8. Check for already reconstructed cases and set the `-time $timeString` argument for the reconstructPar tool
 countRec=0
 realToDoReconstruct[$countRec]=-1
 for ii in ${!arrayReconstruct[@]}; do
    correctReconstruct[$ii]="true" #Initialising this array to be used in the following subsections (9.)
    timeHere=${arrayReconstruct[ii]}
-   if [ -f ${timeHere}/.reconstructDone ]; then
-      echo "Time ${timeHere} has already been reconstructed. No reconstruction will be performed"
+   if [ -f ${caseDir}/${timeHere}/.reconstructDone ]; then
+      echo "Time ${timeHere} has already been reconstructed in ${caseDir}. No reconstruction will be performed"
+   elif [ -f ${reconstructDir}/${timeHere}/.reconstructDone ]; then
+      echo "Time ${timeHere} has already been reconstructed in ${reconstructDir}. No reconstruction will be performed"
    else
       realToDoReconstruct[$countRec]=$timeHere
      (( countRec++ ))
@@ -133,7 +167,7 @@ if [ $countRec -gt 0 ]; then
    unset leftToDoReconstruct
    leftToDoReconstruct=("${realToDoReconstruct[@]}")
    while [ $kStart -lt ${#realToDoReconstruct[@]} ]; do
-      ## 8. Create the batch to reconstruct
+      ## 9. Create the batch to reconstruct
       timeString=""
       countHere=0
       unset hereToDoReconstruct
@@ -172,7 +206,7 @@ if [ $countRec -gt 0 ]; then
          echo "kNext=$kNext"
       fi
 
-      ## 9. Copy from the ./overlayFSDir/overlay* the full batch into ./bakDir/bak.processor*
+      ## 10. Copy from the ./overlayFSDir/overlay* the full batch into ./bakDir/bak.processor*
       unset arrayCopyIntoBak
       arrayCopyIntoBak=("${hereToDoReconstruct[@]}")
       replace="true"
@@ -182,18 +216,18 @@ if [ $countRec -gt 0 ]; then
          echo "Exiting";exit 1
       fi
       
-      ## 10. Reconstruct all times for this batch.
-      echo "Start reconstruction"
+      ## 11. Reconstruct all times for this batch.
+      echo "Start reconstruction of timeString=$timeString"
       logFileHere=$logsDir/log.reconstructPar.${SLURM_JOBID}_${hereToDoReconstruct[0]}-${hereToDoReconstruct[-1]}
       srun -n 1 -N 1 singularity exec $theImage reconstructPar -time ${timeString} 2>&1 | tee $logFileHere
       
-      ## 11. Mark successfully reconstructed times with the "flag" file: .reconstructDone, and remove the decomposed time results
+      ## 12. Mark successfully reconstructed times with the "flag" file: .reconstructDone, and remove the decomposed time results, and move the reconstructed results to the case directory
       noErrors="true"
       for ii in ${!arrayReconstruct[@]}; do
          timeHere=${arrayReconstruct[ii]}
          indexInHereTodo=$(getIndex "${timeHere}" "${hereToDoReconstruct[@]}")
          if [ $indexInHereTodo -ne -1 ]; then
-            ###11.1 Check if reconstruction was successful by looking for errors in the log file
+            ###12.1 Check if reconstruction was successful by looking for errors in the log file
             checkOn="false"
             while IFS= read -r line; do
                if [ "${checkOn}" = "true" ]; then
@@ -217,17 +251,28 @@ if [ $countRec -gt 0 ]; then
                fi
             done < $logFileHere
          
-            ###11.2 If reconstruction was correct, mark the reconstructed folder
+            ###12.2 If reconstruction was correct, mark the reconstructed folder
             if [ "${correctReconstruct[ii]}" = "true" ]; then
-               touch ${timeHere}/.reconstructDone
-               echo "Reconstruction finished, and file ${timeHere}/.reconstructionDone created."
+               if [ -d ${reconstructDir}/${timeHere} ]; then
+                  touch ${reconstructDir}/${timeHere}/.reconstructDone
+                  echo "Reconstruction finished, and file ${timeHere}/.reconstructionDone created."
+                  mv "${reconstructDir}/${timeHere}" "${caseDir}/${timeHere}"
+                  if [ ! -f ${caseDir}/${timeHere}/.reconstructDone ]; then
+                     echo "${caseDir}/${timeHere}/.reconstructDone was not found"
+                     noErrors="false"
+                  fi
+               else
+                  echo "The directory ${reconstructDir}/${timeHere} does not exist"
+                  noErrors="false"
+               fi
             else
                echo "Check why Time ${timeHere} failed in reconstruction"
                noErrors="false"
             fi
          fi
       done
-      ## 12. Remove the decomposed time results if they have been successfully reconstructed (except those indicated to be kept)
+
+      ## 13. Remove the decomposed time results if they have been successfully reconstructed (except those indicated to be kept)
       if [ "${noErrors}" = "true" ]; then
          for ii in ${!arrayReconstruct[@]}; do
             timeHere=${arrayReconstruct[ii]}
@@ -254,6 +299,7 @@ if [ $countRec -gt 0 ]; then
          echo "There were some errors during reconstruction. This script will stop here"
          echo "Please check the log files to investigate the possible sources of reconstruction errors"
          echo "Decomposed results currently in the host will not be removed to assist with the investigation"
+         echo "Stop recursive jobs if they exist"
          echo "Exiting"; exit 1
       fi
       kStart=$kNext
@@ -262,42 +308,50 @@ else
    echo "The times asked to be reconstructed are already reconstructed"
 fi
 
-#13. As a final step, remove any decomposed time result that was left in the host file system, but to be reomoved
+#14. As a final step, remove any decomposed time result that was left in the host file system, but to be removed
 #    (as it was already reconstructed successfully and is not indicated to be kept)
-##13.1 Generating the array of existing decomposed times in the local host
-unset arrayReconstruct #This global variable will be re-created in the function below
-generateReconstructArray "${keepTimesArr[0]}:${keepTimesArr[-1]}" "bak";success=$? #Calling function to generate "arrayReconstruct"
-if [ $success -ne 0 ]; then
-   echo "Failed creating the arrayReconstruct"
-   echo "Exiting";exit 1
-fi
-##13.2 Deleting decomposed results, except those indicated to be kept
-for ii in ${!arrayReconstruct[@]}; do
-   timeHere=${arrayReconstruct[ii]}
-   indexKeep=$(getIndex "${timeHere}" "${keepTimesArr[@]}")
-   if [ $indexKeep -eq -1 ]; then
-      if [ -f ${timeHere}/.reconstructDone ]; then
-         echo "Removing time ${timeHere} in the ./bakDir/bak.processor* directories"
-         for jj in $(seq 0 $((foam_numberOfSubdomains -1))); do
-            if [ -d ./bakDir/bak.processor${jj}/${timeHere} ]; then
-               srun -n 1 -N 1 --mem-per-cpu=0 --exclusive find -P ./bakDir/bak.processor${jj}/${timeHere} -type f -print0 -type l -print0 | xargs -0 munlink &
-            fi
-         done
-         wait
-         for jj in $(seq 0 $((foam_numberOfSubdomains -1))); do
-            if [ -d ./bakDir/bak.processor${jj}/${timeHere} ]; then
-               srun -n 1 -N 1 --mem-per-cpu=0 --exclusive find ./bakDir/bak.processor${jj}/${timeHere} -depth -type d -empty -delete &
-            fi
-         done
-         wait
-      else
-         echo "Cannot remove time ${timeHere} in the ./bakDir/bak.processor* directories"
-         echo "Because the reconstruction is not marked with ${timeHere}/.reconstructDone"
-         echo "Something has gone wrong. Please revise the log files to catch the problem"
-         echo "Exiting"; exit 1
-      fi 
+##14.0 Check if there are times left in ./bakDir/bak.processor* directories (asking for the last one here)
+lastTimeReached=$(getNResultTime -1 "bak");success=$? #Calling function to obtain the Last Time result available (-1)
+echo "lastTimeReached=$lastTimeReached"
+if [ $lastTimeReached -eq -1 ]; then
+   echo "The ./bakDir/bak.processor* directories have no more results saved within"
+   echo "Nothing else to be removed from the file system"
+else
+   ##14.1 Generating the array of existing decomposed times in the local host
+   unset arrayReconstruct #This global variable will be re-created in the function below
+   generateReconstructArray "${keepTimesArr[0]}:${keepTimesArr[-1]}" "bak";success=$? #Calling function to generate "arrayReconstruct"
+   if [ $success -ne 0 ]; then
+      echo "Failed creating the arrayReconstruct"
+      echo "Exiting";exit 1
    fi
-done
+   ##14.2 Deleting decomposed results, except those indicated to be kept
+   for ii in ${!arrayReconstruct[@]}; do
+      timeHere=${arrayReconstruct[ii]}
+      indexKeep=$(getIndex "${timeHere}" "${keepTimesArr[@]}")
+      if [ $indexKeep -eq -1 ]; then
+         if [ -f ${timeHere}/.reconstructDone ]; then
+            echo "Removing time ${timeHere} in the ./bakDir/bak.processor* directories"
+            for jj in $(seq 0 $((foam_numberOfSubdomains -1))); do
+               if [ -d ./bakDir/bak.processor${jj}/${timeHere} ]; then
+                  srun -n 1 -N 1 --mem-per-cpu=0 --exclusive find -P ./bakDir/bak.processor${jj}/${timeHere} -type f -print0 -type l -print0 | xargs -0 munlink &
+               fi
+            done
+            wait
+            for jj in $(seq 0 $((foam_numberOfSubdomains -1))); do
+               if [ -d ./bakDir/bak.processor${jj}/${timeHere} ]; then
+                  srun -n 1 -N 1 --mem-per-cpu=0 --exclusive find ./bakDir/bak.processor${jj}/${timeHere} -depth -type d -empty -delete &
+               fi
+            done
+            wait
+         else
+            echo "Cannot remove time ${timeHere} in the ./bakDir/bak.processor* directories"
+            echo "Because the reconstruction is not marked with ${timeHere}/.reconstructDone"
+            echo "Something has gone wrong. Please revise the log files to catch the problem"
+            echo "Exiting"; exit 1
+         fi 
+      fi
+   done
+fi
 
 #X. Final step
 echo "Script done"
